requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.25.2
python-dotenv==1.0.0
requests==2.31.0